NSIS: rework uninstall logic for LilyPad.
Before this commit, the NSIS installer created three versions of LilyPad
(lilypad.exe, lilypad-ascii.exe, and lilypad-unicode.exe), but we should
only need two of them.

If the underlying OS does not support Unicode, we only need

  lilypad.exe
  lilypad-unicode.exe

and if Unicode *is* supported, then we need

  lilypad.exe
  lilypad-ascii.exe

Then, the default LilyPad is always "lilypad.exe", and we don't have the
extra, duplicated executable.

This commit implements the above logic, and adds "lilypad-unicode.exe"
to the list of installed files if necessary.