tools::python: explicitly set flags for building dbm module
Comment from Jan:
in target/tools/src/python-2.4.5/setup.py:527 (and below):

        db_setup_debug = False   # verbose debug prints from this script?

Setup.py looks for /usr/include/*db*3,4*/db.h, which is entirely broken:
it should ask GCC if it can find db.h (and/or look at the include
flags); we are cross compiling and it entirely ignores that option.

This holds for all of the extension modules that are checked here.  It
would be nice if we could add %(system_prefix)s for where it looks for
all extensions; on the other hand: trusting GUB's dependencies and
patching Setup.dist is OK too.